feat(agent-runtime): add runtime helpers for Copilot LLM execution
