+++ src/qdash/common/copilot/agent_runtime/__init__.py
@@ -1,0 +1,1 @@
+"""Runtime helpers for Copilot LLM agent execution."""
