+++ src/qdash/common/copilot/agent_runtime/schemas.py
@@ -1,0 +1,49 @@
+"""Response schema contracts for Copilot LLM outputs."""
+
+from __future__ import annotations
+
+from typing import Any
+
+ANALYSIS_RESPONSE_SCHEMA: dict[str, Any] = {
+    "type": "object",
+    "properties": {
+        "summary": {"type": "string"},
+        "assessment": {"type": "string", "enum": ["good", "warning", "bad"]},
+        "explanation": {"type": "string"},
+        "potential_issues": {"type": "array", "items": {"type": "string"}},
+        "recommendations": {"type": "array", "items": {"type": "string"}},
+    },
+    "required": [
+        "summary",
+        "assessment",
+        "explanation",
+        "potential_issues",
+        "recommendations",
+    ],
+    "additionalProperties": False,
+}
+
+BLOCKS_RESPONSE_SCHEMA: dict[str, Any] = {
+    "type": "object",
+    "properties": {
+        "blocks": {
+            "type": "array",
+            "items": {
+                "type": "object",
+                "properties": {
+                    "type": {"type": "string", "enum": ["text", "chart"]},
+                    "content": {"type": ["string", "null"]},
+                    "chart": {"type": ["object", "null"]},
+                },
+                "required": ["type", "content", "chart"],
+                "additionalProperties": False,
+            },
+        },
+        "assessment": {
+            "type": ["string", "null"],
+            "enum": ["good", "warning", "bad", None],
+        },
+    },
+    "required": ["blocks", "assessment"],
+    "additionalProperties": False,
+}
