+++ src/qdash/workflow/engine/task/ai_triage.py
@@ -347,7 +347,7 @@ def _run_ai_triage(
     config: CopilotConfig,
 ) -> str | None:
     """Run Copilot analysis and return markdown content."""
-    from qdash.common.copilot.llm_agent import blocks_to_markdown, run_analysis
+    from qdash.common.copilot.agent import blocks_to_markdown, run_analysis
     from qdash.common.copilot.runtime import CopilotRuntime
 
     config = _ai_triage_config(config)
